feat: larger season selection
