+++ apps/web/src/app/teams/[team]/components/team-roster.tsx
@@ -1,13 +1,7 @@
 "use client";
 
-import { useState, useEffect } from "react";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
+import { useState } from "react";
+
 import {
   Table,
   TableBody,
@@ -21,37 +15,22 @@ import { ArrowUp } from "lucide-react";
 import { Button } from "@/components/ui/button";
 
 import { useRoster } from "@/hooks/requests/useRoster";
-import { useSeasons } from "@/hooks/requests/useSeasons";
-import { useRouter, usePathname } from "next/navigation";
 
 type SortField = "name" | "position" | "gender" | "gp" | "goals" | "gpg";
 type SortDirection = "asc" | "desc";
 
-export function TeamRoster({ teamId }: { teamId: string }) {
+export function TeamRoster({
+  seasonId,
+  teamId,
+  seasonYear,
+}: {
+  seasonId: string;
+  teamId: string;
+  seasonYear: string;
+}) {
   const [sortField, setSortField] = useState<SortField>("name");
   const [sortDirection, setSortDirection] = useState<SortDirection>("asc");
-  const [seasonId, setSeasonId] = useState<string>("");
-  const router = useRouter();
-  const pathname = usePathname();
-
-  const { seasons } = useSeasons({});
-
-  useEffect(() => {
-    setSeasonId(
-      seasons?.filter((season: Season) => season.is_current)[0].id
-    );
-  }, []);
-
-  console.log("seasonId:", seasonId);
   const { roster } = useRoster({ seasonId, teamId });
-
-  const handleSeasonChange = (newSeasonId: string) => {
-    // Update the URL with the new season ID
-    const searchParams = new URLSearchParams(window.location.search);
-    searchParams.set("season", newSeasonId);
-    router.push(`${pathname}?${searchParams.toString()}`);
-    setSeasonId(newSeasonId);
-  };
 
   const handleSort = (field: SortField) => {
     if (field === sortField) {
@@ -107,27 +86,7 @@ export function TeamRoster({ teamId }: { teamId: string }) {
     <Card>
       <CardHeader>
         <div className="flex justify-between items-center">
-          <CardTitle>Roster</CardTitle>
-          <div className="w-32">
-            <Select value={seasonId} onValueChange={handleSeasonChange}>
-              <SelectTrigger>
-                <SelectValue placeholder="Select Season" />
-              </SelectTrigger>
-              <SelectContent>
-                {seasons?.map(
-                  (season: {
-                    id: number;
-                    year: string;
-                    is_current: boolean;
-                  }) => (
-                    <SelectItem key={season.id} value={season.id.toString()}>
-                      {season.year}
-                    </SelectItem>
-                  )
-                )}
-              </SelectContent>
-            </Select>
-          </div>
+          <CardTitle>{seasonYear} Roster</CardTitle>
         </div>
       </CardHeader>
       <CardContent>
